requirements.txt
weaviate-client==4.20.1
weaviate-agents==1.3.0
streamlit==1.54.0
python-dotenv==1.2.1
requests==2.32.5
